1、增加用户同步

+++ wxwork-taskjob/src/main/java/com/ucucs/wxwork/model/WxWorkParty.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 @Getter
 @Setter
 public class WxWorkParty {
-    private Long id;
+  private Long id;
 
-    private String name;
+  private String name;
 
-    private String enName;
+  private String enName;
 
-    private Long parentId;
+  private Long parentId;
 
-    private Long order;
+  private Integer order;
 }
